+++ changelog.md
@@ -21,10 +21,12 @@ This is a major update that breaks backwards compatibility. To emphasise that th
   * `POP3->Version`
 * All elements previously marked as deprecated have been removed (e.g. `ReturnPath`)
 * NTLM authentication removed - never worked anyway!
+  * `PHPMailer->Workstation`
+  * `PHPMailer->Realm`
 * `SMTP::authenticate` method signature changed
 * `parseAddresses()` is now static
 * `validateAddress()` is now called statically from `parseAddresses()`
-* Extensive reworking of XOAUTH2, adding support for Google, Yahoo and Microsoft providers, thanks to @sherryl4george
+* Extensive reworking of XOAUTH2, adding support for Google, Yahoo and Microsoft providers in the standard PHPMailer class, thanks to @sherryl4george
 * Fix extra line break in getSentMIMEMessage()
 
 ## Version 5.2.14 (Nov 1st 2015)